% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{AIC_}
\alias{AIC_}
\title{AIC of two numeric vectors}
\usage{
AIC_(x, y, nbK)
}
\arguments{
\item{x, y}{two numeric vectors of same length.}

\item{nbK}{an integer. \code{nbK} is the number of used parameters.
\code{nbK > 1} : returns \code{NA} if not.}
}
\value{
Return the Akaike Information Criterion (\code{AIC}).
}
\description{
Take two numeric vectors and return the Akaike Information
 Criterion (\code{AIC}).
 \code{AIC} is computed from the Residuals Sum of Square
 \code{\link{rss}} and
 the number of used parameters.
}
\details{
None.
}
\keyword{internal}
