% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{amean}
\alias{amean}
\title{Arithmetic mean}
\usage{
amean(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
Return the arithmetic mean of a numeric vector.
}
\description{
Take a numeric vector and return the arithmetic mean.
}
\details{
The function \code{amean()} works as the standard function
 \code{stats::mean()}, but with option \code{na.rm = TRUE} by default.
}
\seealso{
\code{\link{amean}}  returns an arithmetic mean. \cr
\code{\link{gmean}}  returns a geometric mean. \cr
\code{\link{wamean}} returns a weighted arithmetic mean. \cr
\code{\link{wgmean}} returns a weighted geometric mean.
}
\keyword{internal}
