% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{argchr}
\alias{argchr}
\title{Index of different instances of a given char in a string}
\usage{
argchr(str, chr)
}
\arguments{
\item{str}{a string.}

\item{chr}{a char of \code{nchar(chr) = 1}.}
}
\value{
Return a vector of indices
of the different instances of the character \code{chr} in \code{str}.
}
\description{
Take a string \code{str}, then look for
 the index of the different instances of the char \code{chr}.
}
\details{
None.
}
\seealso{
\code{\link{argchr}} returns all the instances
of \code{chr} in \code{str}. \cr
\code{\link{first_argchr}} returns the first instance
of \code{chr} in \code{str}. \cr
\code{\link{delstr_begin}} deletes the beginning of a string. \cr
\code{\link{delstr_end}} deletes the end of a string. \cr
}
\keyword{internal}
