% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{argmax}
\alias{argmax}
\title{Index of the maximum values of a vector}
\usage{
argmax(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
Return the index of the maximum values of a numeric vector.
}
\description{
Take a numeric vector and return the index of its
 maximum values.
}
\details{
The function \code{argmax()} works as the standard function
 \code{which.max()}.
}
\seealso{
\code{\link{argmin}}, \code{\link{first_argmin}},
         \code{\link{first_argmax}}
}
\keyword{internal}
