% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{argmin}
\alias{argmin}
\title{Index of the minimum values of a vector}
\usage{
argmin(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
Return the index of the minimum values of a numeric vector.
}
\description{
Take a numeric vector and return the index of
its minimum values.
}
\details{
The function \code{argmin()} works as the standard
function \code{which.min()}.
}
\seealso{
\code{\link{first_argmin}}, \code{\link{argmax}},
         \code{\link{first_argmax}}
}
\keyword{internal}
