% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{asd}
\alias{asd}
\title{Arithmetic standard deviation}
\usage{
asd(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
Return the arithmetic standard deviation of a numeric vector,
 as the standard function \code{stats::sd()}.
}
\description{
Take a numeric vector
and return its arithmetic standard deviation.
}
\details{
The function \code{asd()} works as the standard function
\code{stats::sd()}, but with option \code{na.rm = TRUE} by default.
}
\seealso{
\code{\link{wasd}} returns arithmetic standard deviation. \cr
\code{\link{wasd}} returns weighted arithmetic standard deviation. \cr
\code{\link{gsd}}  returns geometric standard deviation. \cr
\code{\link{wgsd}} returns weighted geometric standard deviation.
}
\keyword{internal}
