% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{binary_to_logical}
\alias{binary_to_logical}
\title{Convert a binary into logical matrix}
\usage{
binary_to_logical(mat)
}
\arguments{
\item{mat}{a binary/numeric matrix.}
}
\value{
Return a logical matrix of \code{dim(mat)}.
}
\description{
Take the binary matrix \code{mat} and convert it
  in a logical matrix.
}
\details{
None.
}
\seealso{
\code{\link{logical_to_binary}}
}
\keyword{internal}
