% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_fclust.R
\name{build_random_matrix}
\alias{build_random_matrix}
\title{Build a matrix of missing indices}
\usage{
build_random_matrix(nbIter = 1, seed = NULL,
                    rm.number = 0, rm.number.max = 0)
}
\arguments{
\item{nbIter}{an integer, that indicates the number of random drawing to do.}

\item{seed}{an integer, that fixes a seed for random drawing.}

\item{rm.number}{an integer,
that indicates the number of elements to randomly remove.}

\item{rm.number.max}{an integer,
that indicates the maximum number of elements
among which the elements to remove are randomly drawn.}
}
\value{
a matrix of indices, of dimensions \code{nbIter} * \code{rm.number}.
}
\description{
Build a matrix of indices of elements hat will be ommitted,
 randomly drawn among a whole set of elements.
}
\details{
Each line (each set of indices) of the matrix
is unique, different from other lines of the matrix.
}
\keyword{internal}
