% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrating.R
\name{calibrate_gmean_bymot_xpr}
\alias{calibrate_gmean_bymot_xpr}
\title{Modelling of performances
by assembly motif
over several experiments}
\usage{
calibrate_gmean_bymot_xpr(fobs, assMotif, xpr)
}
\arguments{
\item{fobs}{a numeric vector. The vector \code{fobs} contains the
quantitative performances of assemblages.}

\item{assMotif}{a vector of labels of \code{length(fobs)}.
The vector \code{assMotif} contains the assembly motifs of assemblages.}

\item{xpr}{a vector of numerics of \code{length(fobs)}.
The vector \code{xpr} contains the weight of each experiment,
and the labels (in \code{names(xpr)}) of different experiments.
The weigth of each experiment is used
in the computation of the Residual Sum of Squares
in the function \code{rss_clustering}.
All assemblages that belong to a given experiment
should then have a same weigth.
Each experiment is identified by its names (\code{names(xpr)})
and the RSS of each experiment is weighted by values of \code{xpr}.
The vector \code{xpr} is generated
by the function \code{stats::setNames}.}
}
\value{
Return a vector of \code{length(fobs)}.
Its values are computed as the geometric mean
of performances of assemblages
that share a same assembly motif,
experiment by experiment.
}
\description{
Take a vector \code{fobs} of assembly performances
over several experiments
and return a vector of performances
predicted as the geometric mean of performances
of all assemblages that share the same assembly motif.\cr

Assembly motifs are labelled in the vector \code{assMotif}.
Experiments are labelled in the vector \code{xpr}.
}
\details{
Modelled performances are computed
using geometric mean (\code{opt.mean = "gmean"})
of performances of assemblages
that share a same assembly motif (\code{opt.model = "bymot"}),
experiment by experiment,
by including all assemblages that belong to a same assembly motif.
}
\seealso{
\code{\link{calibrate_amean_bymot_xpr}} arithmetic mean
  by assembly motif. \cr
\code{\link{calibrate_amean_byelt_xpr}} arithmetic mean
  by elements occurring within assembly motif. \cr
\code{\link{calibrate_gmean_bymot_xpr}} geometric mean
  by assembly motif. \cr
\code{\link{calibrate_gmean_byelt_xpr}} geometric mean
  by elements occurring within assembly motif. \cr
}
\keyword{internal}
