% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CedarCreek.2004.2006.boot.assemblages}
\alias{CedarCreek.2004.2006.boot.assemblages}
\title{Evaluate by bootstrapping the robustness of species clustering
to the number of assemblages (plots) used in the cluster analysis.}
\format{A list of matrices, each matrix containing
 the results for a given clustering index. The indices are
 "Czekanowski_Dice", "Folkes_Mallows", "Jaccard", "Kulczynski",
 "Precision", "Rand", "Recall", "Rogers_Tanimoto", "Russel_Rao",
 "Sokal_Sneath1" and "Sokal_Sneath2" index.}
\usage{
CedarCreek.2004.2006.boot.assemblages
}
\description{
Species assemblages are randomly removed from the dataset,
from 1 to m-1 assemblages (with m = total number of assemblages),
a functional clustering of species is built,
and the effect of this assemblage remove on species clustering
is evaluated by comparing the resulting species clustering
with the species clustering obtained with all observed species assemblages.
The perturbation is measured by using different clustering indices.
}
\details{
None.
}
\keyword{datasets}
