% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CedarCreek.2004.2006.boot.performances}
\alias{CedarCreek.2004.2006.boot.performances}
\title{Evaluate by bootstrapping the robustness of species clustering
to the number of performances (annual biomass) used in the cluster analysis.}
\format{A list of matrices, each matrix containing
 the results for a given clustering index. The indices are
 "Czekanowski_Dice", "Folkes_Mallows", "Jaccard", "Kulczynski",
 "Precision", "Rand", "Recall", "Rogers_Tanimoto", "Russel_Rao",
 "Sokal_Sneath1" and "Sokal_Sneath2" index.}
\usage{
CedarCreek.2004.2006.boot.performances
}
\description{
Performances are randomly removed from the dataset,
from 1 to n-1 performances (with n = total number of performances),
a functional clustering of species is built,
and the effect of this performance remove on species clustering
is evaluated by comparing the resulting species clustering
with the species clustering obtained with all systemic performances.
The perturbation is measured by using different clustering indices.

Here, the total number of performances is 3 only,
then the interest of this process is very limited.
}
\details{
None.
}
\keyword{datasets}
