% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CedarCreek.2004.2006.test.performances}
\alias{CedarCreek.2004.2006.test.performances}
\title{Test of significance of different performances (yearly biomass)
used in Cedar Creek experiment for the years 2004, 2005 and 2006}
\format{A list of matrices, each matrix containing
 the results for a given clustering index. The indices are
 "Czekanowski_Dice", "Folkes_Mallows", "Jaccard", "Kulczynski",
 "Precision", "Rand", "Recall", "Rogers_Tanimoto", "Russel_Rao",
 "Sokal_Sneath1" and "Sokal_Sneath2" index.}
\usage{
CedarCreek.2004.2006.test.performances
}
\description{
Each performance is successively removed
from the data set, a functional clustering of species is built,
and the effect of this performance remove on species clustering
is evaluated by comparing the resulting species clustering
with the species clustering obtained with all systemic performances.
The perturbation is measured by using different clustering indices.
}
\details{
None.
}
\keyword{datasets}
