% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CedarCreek.2004.res}
\alias{CedarCreek.2004.res}
\title{Functional clustering of species used in Cedar Creek experiment
for the year 2004}
\format{A hierarchical tree of species clustering:
\describe{
\item{tree$aff}{a square-matrix of dimensions species number x
species number}
\item{tree$cor}{a vector of coefficient of determination}
}}
\usage{
CedarCreek.2004.res
}
\description{
Cedar Creek results obtained for the biomass production
in the year 2004 only.
}
\details{
None.
}
\keyword{datasets}
