% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{char_to_int}
\alias{char_to_int}
\title{Convert a character into integer vector}
\usage{
char_to_int(v)
}
\arguments{
\item{v}{a character vector.}
}
\value{
Return a integer vector.
}
\description{
Take the charactervector \code{v} and convert it
  in an integer vector.
}
\details{
None.
}
\keyword{internal}
