% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{check_foption}
\alias{check_foption}
\title{Check an option}
\usage{
check_foption(opt, all)
}
\arguments{
\item{opt}{a string, that is a possible otion.}

\item{all}{a vector of strings, that iS a set of possible options.}
}
\value{
Stop the program, or return a string which is a right option.
}
\description{
Take a string which is an option,
 check if the string is among a given set of options,
 then stop the program if option is wrong,
 or return the right and complete option.
}
\details{
None.
}
\keyword{internal}
