% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{check_ftree}
\alias{check_ftree}
\title{Check a matrix of component affectation to functional groups}
\usage{
check_ftree(mat)
}
\arguments{
\item{mat}{a square-matrix of integer.}
}
\value{
Return a regular matrix.
}
\description{
Take a matrix of component affectation,
then check that the numero of each affectation
increases regularly,
from \code{1} until to the number of components,
with the number of functional groups.
}
\details{
None.
}
\keyword{internal}
