% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fclust.R
\name{check_plot_options}
\alias{check_plot_options}
\title{Check options}
\usage{
check_plot_options(opt, all, main = "", windowmin = NULL,
  affectElt = NULL)
}
\arguments{
\item{opt}{a list of options for a considered plot function}

\item{all}{a list that contains all possible options
for the  considered plot function}

\item{main}{the name of plot function}

\item{windowmin}{a vector of indices}

\item{affectElt}{a vector of integer.}
}
\value{
a list of consistant options
}
\description{
Check the consistence of options for plotting fclust-results
}
\details{
None.
}
\keyword{internal}
