% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{concat_by_line}
\alias{concat_by_line}
\title{Concat a list of strings segmented by line}
\usage{
concat_by_line(str, sepChar = ",", nbChar = 35)
}
\arguments{
\item{str}{a list of strings.}

\item{sepChar}{a string.
The string is a separator between each element of list of strings.}

\item{nbChar}{an integer.
It is the length of each line of strings.}
}
\value{
Return a vector of strings,
the length of each string being lower than \code{nbChar}.
}
\description{
Take a list of strings,
concat all the strings in a single long string
by separating each one by the given separator \code{sepChar},
then segment this long string into strings of a given length \code{nChar}.
}
\details{
None.
}
\keyword{internal}
