% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{delstr_begin}
\alias{delstr_begin}
\title{Delete the beginning of a string}
\usage{
delstr_begin(str, n)
}
\arguments{
\item{str}{a string of length \code{nchar(str)}.}

\item{n}{an integer.
It is the number of characters to delete
at the beginning of the string \code{str}}
}
\value{
A string of length \code{nchar(str) - n},
shortened by its beginning.
}
\description{
Take a string, then delete the begining of this string.
}
\details{
None.
}
\seealso{
\code{\link{argchr}} returns all the instances
of \code{chr} in \code{str}. \cr
\code{\link{first_argchr}} returns the first instance
of \code{chr} in \code{str}. \cr
\code{\link{delstr_begin}} deletes the beginning of a string. \cr
\code{\link{delstr_end}} deletes the end of a string. \cr
}
\keyword{internal}
