% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{extend_letters}
\alias{extend_letters}
\title{Extends a vector of letters}
\usage{
extend_letters(long)
}
\arguments{
\item{long}{an integer, that corresponds
to the whished \code{length(vector) = long}.}
}
\value{
Return a vector of \code{length(vector) = long}
built by concatening lowercase and uppercase letters with an integer,
until \code{length(vector) = long}.
}
\description{
Take the vector of lowercase and uppercase letters,
then extend it until to the length of vector equal to \code{long}.
}
\details{
None.
}
\keyword{internal}
