% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{extend_vector}
\alias{extend_vector}
\title{Extends a vector}
\usage{
extend_vector(x, long)
}
\arguments{
\item{x}{any vector.}

\item{long}{an integer, that corresponds
to the whished \code{length(x) = long}.}
}
\value{
Return a vector of \code{length(x) = long}
built by recycling \code{x} as many times as necessary,
then truncated at \code{length(x) = long}.
}
\description{
Take the vector \code{x},
then extend it until to \code{length(x) = long}.
}
\details{
None.
}
\keyword{internal}
