% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.R
\name{fclust_predict}
\alias{fclust_predict}
\title{nnnn}
\usage{
fclust_predict(tree, res.prd, supOccur, appOccur, appFct,
  opt.mean = "amean", opt.model = "byelt")
}
\arguments{
\item{tree}{vv}

\item{res.prd}{ff}

\item{supOccur}{ff}

\item{appOccur}{ff}

\item{appFct}{ff}

\item{opt.mean}{ff}

\item{opt.model}{ff}
}
\value{
vv
}
\description{
vvv
}
\details{
ccc
}
\keyword{internal}
