% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{fcolours}
\alias{fcolours}
\title{Series of colours used in \pkg{functClust}}
\usage{
fcolours()
}
\value{
Return the vector of the colours
 used in the package \pkg{functClust}.
}
\description{
Return the colours used in the package \pkg{functClust}.
}
\details{
The vector of colours is
\code{c("red3", "blue2", "orange2", "turquoise3", "magenta", "green4")}.\cr

 The use of functions \code{fsymbols() * fcolours()} generates
 a vector of \code{5 * 6 = 30} different symbols,
 before to be automatically recycled by R.
}
\seealso{
\code{\link{fepsilon}} returns a value for epsilon. \cr
\code{\link{fletters}} returns a vector of lowercase
then uppercase letters. \cr
\code{\link{fcolours}} returns a vector of colours. \cr
\code{\link{fsymbols}} returns a vector of symbols.
}
\keyword{internal}
