% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{fepsilon}
\alias{fepsilon}
\title{Value of epsilon used in \pkg{functClust}}
\usage{
fepsilon()
}
\value{
Return the constant value of epsilon
}
\description{
Return the constant value of epsilon.
}
\details{
Epsilon is a constant, equals to \code{1e-12}.
}
\seealso{
\code{\link{fepsilon}} returns a value for epsilon. \cr
\code{\link{fletters}} returns a vector of lowercase
  then uppercase letters. \cr
\code{\link{fcolours}} returns a vector of colours. \cr
\code{\link{fsymbols}} returns a vector of fsymbols.
}
\keyword{internal}
