% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{first_argchr}
\alias{first_argchr}
\title{Index of a first instance of char in a string}
\usage{
first_argchr(str, chr)
}
\arguments{
\item{str}{a string.}

\item{chr}{a char of \code{length(chr) = 1}.}
}
\value{
Return the position index of the first \code{chr} in \code{str}.
}
\description{
Take a string \code{str}, then look for
 the position index of the first instance of the char \code{chr}.
}
\details{
None.
}
\seealso{
\code{\link{argchr}} returns all the instances
of \code{chr} in \code{str}. \cr
\code{\link{first_argchr}} returns the first instance
of \code{chr} in \code{str}. \cr
\code{\link{delstr_begin}} deletes the beginning of a string. \cr
\code{\link{delstr_end}} deletes the end of a string. \cr
}
\keyword{internal}
