% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{fletters}
\alias{fletters}
\title{Series of letters used in \pkg{functClust}}
\usage{
fletters()
}
\value{
Return the vector of letters
 used in the package \pkg{functClust}.
}
\description{
Concatenate lowercase and uppercase letters
used in the package \pkg{functClust}.
}
\details{
The vector of letters is \code{c(letters, LETTERS)},
which corresponds to lowercase letters, followed by uppercase letters.
}
\seealso{
\code{\link{fepsilon}} returns a value for epsilon. \cr
\code{\link{fletters}} returns a vector of lowercase
then uppercase letters. \cr
\code{\link{fcolours}} returns a vector of colours. \cr
\code{\link{fsymbols}} returns a vector of fsymbols.
}
\keyword{internal}
