% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{fnbdigits}
\alias{fnbdigits}
\title{Number of digits kept when writing in a file}
\usage{
fnbdigits()
}
\value{
Return the number of digits kept when writing in a file.
}
\description{
Return the number of digits kept when writing in a file.
}
\details{
None.
}
\seealso{
\code{\link{fepsilon}}: return a value for epsilon. \cr
}
\keyword{internal}
