% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{fpvalue}
\alias{fpvalue}
\title{Standard p-value used in the package functClust}
\usage{
fpvalue()
}
\value{
Return the p-value used in the package functClust.
}
\description{
Return the p-value used in the package functClust
}
\details{
None.
}
\seealso{
\code{\link{fepsilon}}: return a value for epsilon. \cr
}
\keyword{internal}
