% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{fstd_colour}
\alias{fstd_colour}
\title{Standard colour used in the package functClust}
\usage{
fstd_colour()
}
\value{
Return the standard colour used in the package functClust.
}
\description{
Return the standard colour used in the package functClust
}
\details{
None.
}
\seealso{
\code{\link{fepsilon}}: return a value for epsilon. \cr
}
\keyword{internal}
