% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{fsymbols}
\alias{fsymbols}
\title{Series of symbols used in \pkg{functClust}}
\usage{
fsymbols()
}
\value{
Return a vector of symbols
 used in the package \pkg{functClust}.
}
\description{
Return the symbols used in the package \pkg{functClust}.
}
\details{
The vector of symbols is \code{c(21, 22, 24, 23, 25)}.
It corresponds to: \code{21 = "circle"}, \code{22 = "square"},
\code{23 = "diamond"}, \code{24 = "triangle point-up"},
\code{25 = "triangle point-down"}. \cr

 The use of functions \code{fsymbols() * fcolours()} generates
 a vector of \code{5 * 6 = 30} different symbols.
}
\seealso{
\code{\link{fepsilon}} returns a value for epsilon. \cr
\code{\link{fletters}} returns a vector of lowercase
then uppercase letters. \cr
\code{\link{fcolours}} returns a vector of colours. \cr
\code{\link{fsymbols}} returns a vector of symbols.
}
\keyword{internal}
