% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{fwindow}
\alias{fwindow}
\title{Standard number of components to plot in the package functClust}
\usage{
fwindow()
}
\value{
Return the standard number of components
to plot in the package functClust.
}
\description{
Return the standard number of components
to plot in the package functClust
}
\details{
None.
}
\seealso{
\code{\link{fepsilon}}: return a value for epsilon. \cr
}
\keyword{internal}
