% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{gmean}
\alias{gmean}
\title{Geometric mean}
\usage{
gmean(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
Return the geometric mean of a numeric vector.
}
\description{
Take a numeric vector and return its geometric mean.
}
\details{
None.
}
\seealso{
\code{\link{amean}} returns an arithmetic mean. \cr
\code{\link{gmean}} returns a geometric mean. \cr
\code{\link{wamean}} returns a weighted arithmetic mean. \cr
\code{\link{wgmean}} returns a weighted geometric mean.
}
\keyword{internal}
