% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{gsd}
\alias{gsd}
\title{Geometric standard deviation}
\usage{
gsd(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
Return the geometric standard deviation of a numeric vector.
}
\description{
Take a numeric vector
 and return its geometric standard deviation.
}
\details{
None.
}
\seealso{
\code{\link{wasd}} returns arithmetic standard deviation. \cr
\code{\link{wasd}} returns weighted arithmetic standard deviation. \cr
\code{\link{gsd}}  returns geometric standard deviation. \cr
\code{\link{wgsd}} returns weighted geometric standard deviation.
}
\keyword{internal}
