% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{index_inturn}
\alias{index_inturn}
\title{Reverse an indexation}
\usage{
index_inturn(index)
}
\arguments{
\item{index}{a vector of integers.}
}
\value{
Return a vector of integers that allows
to find in return the initial order.
}
\description{
Take a vector of indices and return the vector of indices
of \code{length(index)} that allows to find in return the initial order.
}
\details{
None.
}
\keyword{internal}
