% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is_binary}
\alias{is_binary}
\title{Test if a vector is binary}
\usage{
is_binary(mat)
}
\arguments{
\item{mat}{a binary/numeric matrix.}
}
\value{
Return \code{TRUE} is the matrix is a binary matrix.
}
\description{
Take a vector or a matrix,
  then check if all elements are equal to "0 or "1".
}
\details{
None.
}
\keyword{internal}
