% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{list_in_quote}
\alias{list_in_quote}
\title{Concat a list of strings}
\usage{
list_in_quote(str, sepChar = ",")
}
\arguments{
\item{str}{a list or vector of strings or numeric.}

\item{sepChar}{a string.
The string is a separator between each element of list of strings.}
}
\value{
Return a vector of strings between quoted and separated by commas.
}
\description{
Take a list of strings,
quotes each string,
then concat all quoted strings in a single long string
by separating each one by the given separator \code{sepChar},
then segment this long string into strings of a given length \code{nChar}.
}
\details{
None.
}
\keyword{internal}
