% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{logical_to_binary}
\alias{logical_to_binary}
\title{Convert a logical into binary matrix}
\usage{
logical_to_binary(mat)
}
\arguments{
\item{mat}{a logical matrix.}
}
\value{
Return a binary matrix of \code{dim(mat)}.
}
\description{
Take the logical matrix \code{mat} and convert it
    in a binary matrix.
}
\details{
None.
}
\seealso{
\code{\link{binary_to_logical}}
}
\keyword{internal}
