% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fclust.R
\name{mark_string}
\alias{mark_string}
\title{Mathematical mark of an ordered string vector}
\usage{
mark_string(vNum, vStr)
}
\arguments{
\item{vNum}{an ordered (increasing or decreasing) vector of numerics.}

\item{vStr}{the corresponding ordered vector of strings.}
}
\value{
The vector of strings with added mathematical symbol.
}
\description{
Take two vectors: a ordered vector of numeric
and the corresponding of strings. Then compare each term
of the ordered suit of numeric and add "=", ">" or "<"
at the end of each string.
}
\details{
None.
}
\keyword{internal}
