% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mean_fct}
\alias{mean_fct}
\title{Switch mean function.}
\usage{
mean_fct(x, opt.mean = c("amean", "gmean"))
}
\arguments{
\item{x}{a numeric vector.}

\item{opt.mean}{equals to \code{"amean"} or \code{"gmean"}
according to that mean value must be computed with arithmetic
or geometric formula, respectively. There is no default value.}
}
\value{
Return the arithmetic or geometric mean
of the numeric vector \code{x}
according to \code{opt.mean = "amean"} or
\code{opt.mean = "gmean"}.
}
\description{
Switch for arithmetic or geometric mean according
to \code{opt.mean = "amean"} or \code{opt.mean = "gmean"}, respectively.
}
\details{
None.
}
\keyword{internal}
