% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mse}
\alias{mse}
\title{Mean Square Error (MSE).}
\usage{
mse(x, y)
}
\arguments{
\item{x, y}{numeric vectors of same length.}
}
\value{
Return the Mean Square Error (\code{MSE}).
}
\description{
Take two numeric vectors
and return their Mean Square Error (\code{MSE}).
}
\details{
None.
}
\seealso{
\code{\link{R2}}, \code{\link{rss}}, \code{\link{rmse}},
         \code{\link{R2mse}}, \code{\link{pmse}}
}
\keyword{internal}
