% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{nb_tests}
\alias{nb_tests}
\title{Maximal number of clustering models to test}
\usage{
nb_tests(n)
}
\arguments{
\item{n}{an interger, that is the number of components
that belong to the system.}
}
\value{
Return the maximal number of clustering models to test.
}
\description{
Take the number of componens,
then compute the maximal number of clustering models to test.
}
\details{
None.
}
\keyword{internal}
