% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{notify_fclust}
\alias{notify_fclust}
\title{Message for following functional clustering computation}
\usage{
notify_fclust(where, all)
}
\arguments{
\item{where}{an integer, that indicates the last tree level done.}

\item{all}{an integer, that indicates the number of tree levels to do.}
}
\value{
Noting. It is a procedure.
}
\description{
Print a message indicating
  where the functional clustering computation is done.
}
\details{
None.
}
\keyword{internal}
