% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fclust.R
\name{plot_by_page}
\alias{plot_by_page}
\title{Plot a list of strings on several pages}
\usage{
plot_by_page(tab, nbline = 25)
}
\arguments{
\item{tab}{a list of strings to plot.}

\item{nbline}{an integer,
that corresponds to a number of lines to plot by page.}
}
\value{
Nothing. It is a procedure.
}
\description{
Take a list of strings,
then divide it in several units of same length a a page.
A page is defined by a number of lines.
}
\details{
None.
}
\keyword{internal}
