% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fclust.R
\name{plot_clusters_content}
\alias{plot_clusters_content}
\title{Write the components belonging to each functional group}
\usage{
plot_clusters_content(affectElt,
                      opt.sort = TRUE,
                      sepChar  = ", ",
                      nbline   = 25)
}
\arguments{
\item{affectElt}{a vector of integers.
The vector contains the labels of the different functional groups
to which each component belongs.
Each functional group is labelled as an integer.}

\item{opt.sort}{a logical.
If \code{opt.sort = TRUE}, the names of assemblages
are sorted alphabetically inside each motif.}

\item{sepChar}{a string.
The string is a separator between each element of list of strings.}

\item{nbline}{an integer,
that corresponds to a number of lines to plot by page.}
}
\value{
Nothing. It is a procedure.
}
\description{
Write the components belonging to each functional group,
on several pages if need.
}
\details{
None.
}
\keyword{internal}
