% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fclust.R
\name{plot_components_content}
\alias{plot_components_content}
\title{Write the components belonging to each functional group}
\usage{
plot_components_content(mOccur,
                        ordre   = NULL,
                        sepChar = ", ",
                        nbline  = 25 )
}
\arguments{
\item{mOccur}{a matrix of occurrence of components within the assemblages.
\code{dim(mOccur)[1] = length(fobs)}.
\code{dim(mOccur)[2]} equals to the total number of components.}

\item{ordre}{a logical.}

\item{sepChar}{a string.
The string is a separator between each element of list of strings.}

\item{nbline}{an integer,
that corresponds to a number of lines to plot by page.}
}
\value{
Nothing. It is a procedure.
}
\description{
Write the components belonging to each functional group,
on several pages if need.
}
\details{
None.
}
\keyword{internal}
