% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fclust.R
\name{plot_motifs_content}
\alias{plot_motifs_content}
\title{Write the assemblages belonging to each assembly motif}
\usage{
plot_motifs_content(fobs, assMotif, assNames,
                    opt.sort = TRUE,
                    sepChar  = ", ",
                    nbline   = 25  )
}
\arguments{
\item{fobs}{numeric vector.
The vector contains assemblage performances.}

\item{assMotif}{an integer vector of \code{length(fobs)}.
The vector contains the labels of each assemblages.
The labels are used to sorted the assemblages
that share a same assembly motif.}

\item{assNames}{a vector of strings of \code{length(fobs)}.
The vector contains the names of each assemblage.}

\item{opt.sort}{a logical.
If \code{opt.sort = TRUE}, the names of assemblages
are sorted alphabetically inside each motif.}

\item{sepChar}{a string.
The string is a separator between each element of list of strings.}

\item{nbline}{an integer.
The number of lines to plot.}
}
\value{
Nothing. It is a procedure.
}
\description{
Write the assemblages belonging to each assembly motif
on several pages.
}
\details{
None.
}
\keyword{internal}
