% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fclust.R
\name{plot_stats}
\alias{plot_stats}
\title{Plot statistics of combinatorial analysis}
\usage{
plot_stats(mStats, main = "")
}
\arguments{
\item{mStats}{a matrix
that contains all statistics for combinatorial analysis.
The matrix is generated by the function \code{\link{compute_fit_stats}}.}

\item{main}{a string. The string is used as graph title.}
}
\value{
Nothing. It is a procedure.
}
\description{
Take a matrix of statistical results
 and plot Coefficient of Determination (\code{R2}),
 Efficiency (\code{E})
 proportion of missing values,
 and \code{AIc} and \code{AICc},
 \emph{versus} the number of clusters of components.
}
\details{
The colnames of matrix are:
\code{"missing"}, \code{"R2cal"}, \code{"R2prd"}, \code{"pcal"},
\code{"pprd"}, \code{"AIC"}, \code{"AICc"}.
}
\keyword{internal}
