% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pmse}
\alias{pmse}
\title{Probability associated with the Coefficient of determination (R2)}
\usage{
pmse(fprd, fobs, nbK)
}
\arguments{
\item{fprd, fobs}{two numeric vectors.
The first vector contains predicted or modelled values.
The second vector contains the reference (observed) values.}

\item{nbK}{an integer. \code{nbK > 1} : returns an
error if not.}
}
\value{
Return the Probability associated with the Coefficient
of Ddtermination (R2).
}
\description{
Take two numeric vectors and return the
 Probability associated with their Coefficient of determination (R2).
}
\details{
Be careful, the function \code{\link{pmse}} is not symmetrical.
The first argument \code{fprd} is vector of estimated, modelled
or predicted values, the second argument \code{fobs} is
the vector of reference (observed values). Both the vectors have the
same length: \code{length(fprd) = length(fobs)}.
}
\seealso{
\code{\link{R2}}, \code{\link{rss}}, \code{\link{mse}},
         \code{\link{rmse}}, \code{\link{R2mse}}
}
\keyword{internal}
