% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fclust.R
\name{points_sd}
\alias{points_sd}
\title{Plot a point with x- and y-error bars}
\usage{
points_sd(x, y, opt = c("mean", "mean"), fig, col, bg, cex = 2, name = "")
}
\arguments{
\item{x, y}{two numeric vectors, x in abscissa and y in coordinate.}

\item{opt}{two strings, equal to \code{"mean"} or \code{"median"}.
Indicate if the point should be plotted at \code{mean(x), mean(y)},
\code{mean(x), median(y)}, \code{median(x), mean(y)}
or \code{median(x), median(y)}.}

\item{fig, col}{two integers, indicating the symbol and the colour to use.}

\item{bg}{a string, indicating the colour background to use.}

\item{cex}{a numeric, indicating the size of symbol to plot.}

\item{name}{a string, indicating the label to add near the point.}
}
\value{
Nothing. It is a procedure.
}
\description{
Take two vectors,
  compute their mean and standard deviation,
  then plot the mean point with error bars along x- and y-axis.
}
\details{
None.
}
\keyword{internal}
