% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.R
\name{predict_amean_bymot}
\alias{predict_amean_bymot}
\title{Prediction of supplementary assemblages by motif}
\usage{
predict_amean_bymot(appFct, appMotifs, supMotifs)
}
\arguments{
\item{appFct}{a vector of numeric values (assembly properties).}

\item{appMotifs}{a vector of labels of \code{length(fct)} (assembly motifs).}

\item{supMotifs}{a vector of labels of \code{length(fct)} (assembly motifs).}
}
\value{
Return a vector of \code{length(supMotifs)}. The values are
 computed using arithmetic mean
 of components belonging to \code{appMotifs}
 and sharing a same motif.
}
\description{
Takes a numeric f.
}
\details{
Prediction ...
}
\keyword{internal}
