% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting.R
\name{predict_gmean_byelt}
\alias{predict_gmean_byelt}
\title{Prediction of supplementary assemblages computed
     gmean = by using geometric mean
     byelt = by motif WITH taking into account species contribution
     by including all the assemblages, even the one to predict
     for any Function (for instance Fobs)}
\usage{
predict_gmean_byelt(appFct, appMotifs, appOccur,
                        supMotifs, supOccur )
}
\arguments{
\item{appFct}{a vector of numeric values (assembly properties).}

\item{appMotifs}{a vector of labels of \code{length(fct)} (assembly motifs).}

\item{appOccur}{a matrix of occurrence (occurrence of components).
Its first dimension equals to \code{length(fct)}. Its second dimension
 equals to the number of components.}

\item{supMotifs}{a vector of labels of \code{length(fct)} (assembly motifs).}

\item{supOccur}{a matrix of occurrence (occurrence of components).
Its first dimension equals to \code{length(fct)}. Its second dimension
 equals to the number of components.}
}
\value{
cccc
}
\description{
Takes a numeric f.
}
\details{
dd
}
\keyword{internal}
