% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pvalue_dependent_R2mse}
\alias{pvalue_dependent_R2mse}
\title{Test for the dependence of two R2}
\usage{
pvalue_dependent_R2mse(mprd, fobs)
}
\arguments{
\item{mprd}{a matrix of numeric vectors.}

\item{fobs}{the reference vector of length equals to \code{dim(mprd)[2]}.}
}
\value{
Return a vector of p-value of length equals
 to \code{dim(mprd)[1]}.
}
\description{
Test for the dependence of two R2.
}
\details{
The function compare R2 obtained by the models
 \code{mprd[i, ] ~ fobs}.
}
\seealso{
\code{\link{test_dependent_R2}},
         \code{\link{test_dependent_R2mse}}
}
\keyword{internal}
