% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{R2mse}
\alias{R2mse}
\title{Coefficient of Determination (\code{R2})}
\usage{
R2mse(fprd, fobs)
}
\arguments{
\item{fprd, fobs}{two numeric vectors.
The first vector contains predicted or modelled values.
The second vector contains the reference (observed) values.}
}
\value{
Return the Coefficient of Determination (\code{R2}).
}
\description{
Take two numeric vectors
and return their Coefficient of determination (\code{R2}).
}
\details{
Be careful, the function \code{\link{R2mse}} is not symmetrical.
The first argument \code{fprd} is vector of estimated, modelled or
predicted values. The second argument \code{fobs} is the vector
of reference. Both the vectors have the same length:
 \code{length(fprd) = length(fobs)}.
}
\seealso{
\code{\link{R2}}, \code{\link{rss}}, \code{\link{mse}},
         \code{\link{rmse}}, \code{\link{pmse}}
}
\keyword{internal}
