% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{reverse_table}
\alias{reverse_table}
\title{Reverse a table along its first dimension}
\usage{
reverse_table(mat)
}
\arguments{
\item{mat}{a table or a matrix.}
}
\value{
Return a reversed table or matrix.
}
\description{
Take a table, then reverse it along is first dimension,
that is by moving the first rows as the last ones,
and the last rows as the first ones.
The function is useful for re-ordering
an decreasing table into an increasing table, or the inverse.
}
\details{
None.
}
\keyword{internal}
