% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rmse}
\alias{rmse}
\title{Root Mean Square Error (\code{RMSE})}
\usage{
rmse(x, y)
}
\arguments{
\item{x, y}{numeric vectors of same length.}
}
\value{
Return the Root Mean Square Error (\code{RMSE}).
}
\description{
Take two numeric vectors
and return their Root Mean Square Error (\code{RMSE}).
}
\details{
None.
}
\seealso{
\code{\link{R2}}, \code{\link{rss}}, \code{\link{mse}},
         \code{\link{R2mse}}, \code{\link{pmse}}
}
\keyword{internal}
