% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rss}
\alias{rss}
\title{Residual Sum of Squares (\code{RSS}).}
\usage{
rss(x, y)
}
\arguments{
\item{x, y}{two numeric vectors of same length.}
}
\value{
Return the Residual Sum of Square (RSS).
}
\description{
Take two numeric vectors
and return their Residual Sum of Square (RSS).
}
\details{
None.
}
\seealso{
\code{\link{R2}}, \code{\link{mse}}, \code{\link{rmse}},
         \code{\link{R2mse}}, \code{\link{pmse}}
}
\keyword{internal}
