% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fclust.R
\name{sort_matrix}
\alias{sort_matrix}
\title{sort a matrix}
\usage{
sort_matrix(mat)
}
\arguments{
\item{mat}{a numeric matrix}
}
\value{
A vector of indices for properly sort the columns of matrix
}
\description{
Take a matrix, then sorted the colums by increasing values
of 1st, 2nd, 3rd,..    n lines of matrix.
}
\details{
None.
}
\keyword{internal}
