% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{stirling}
\alias{stirling}
\title{Number of Stirling of second kind}
\usage{
stirling(n)
}
\arguments{
\item{n}{a positive integer.
This corresponds to the size of set,
that is the number of components that belong to the set.}
}
\value{
Return a vector of integer. The \code{names} are the number
 \code{k} of clusters. The \code{values} are the number of partitions
  into \code{k} clusters possibly combinated with \code{n} components.
}
\description{
Take an integer and return the number
 of Stirling of second kind, that is the number
 of all possible partitions into \code{k} clusters
  among \code{n} components.
}
\details{
None.
}
\keyword{internal}
