% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{test_dependent_R2mse}
\alias{test_dependent_R2mse}
\title{Test for the dependence of two Coefficients of Determination R2}
\usage{
test_dependent_R2mse(fprd2, fprd1, fobs)
}
\arguments{
\item{fprd2, fprd1, fobs}{three numeric vectors of same length.}
}
\value{
Return a p-value.
}
\description{
Test for the dependence of two R2.
}
\details{
Be careful, the three vectors are not symmetrical.
The function compare R2 obtained by the models \code{fprd1 ~ fobs}
and \code{fprd2 ~ fobs}.
}
\seealso{
\code{\link{test_dependent_R2}},
         \code{\link{pvalue_dependent_R2mse}}
}
\keyword{internal}
