% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{wamean}
\alias{wamean}
\title{Weighted arithmetical mean}
\usage{
wamean(x, w)
}
\arguments{
\item{x}{a numeric vector.}

\item{w}{a vector of numeric weights, of \code{length(x)}.}
}
\value{
Return the weighted arithmetic mean of a numeric vector.
}
\description{
Take a numeric vector and
return the weighted arithmetic mean.
}
\details{
None.
}
\seealso{
\code{\link{amean}}  returns an arithmetic mean. \cr
\code{\link{gmean}}  returns a geometric mean. \cr
\code{\link{wamean}} returns a weighted arithmetic mean. \cr
\code{\link{wgmean}} returns a weighted geometric mean.
}
\keyword{internal}
