% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{wgmean}
\alias{wgmean}
\title{Weighted geometric mean}
\usage{
wgmean(x, w)
}
\arguments{
\item{x}{a numeric vector.}

\item{w}{a vector of weights, of \code{length(x)}.}
}
\value{
Return the weighted geometric standard deviation
of a numeric vector.
}
\description{
Take a numeric vector and return its weighted geometric mean.
}
\details{
None.
}
\seealso{
\code{\link{amean}} returns an arithmetic mean. \cr
\code{\link{gmean}} returns a geometric mean. \cr
\code{\link{wamean}} returns a weighted arithmetic mean. \cr
\code{\link{wgmean}} returns a weighted geometric mean.
}
\keyword{internal}
