% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{wgsd}
\alias{wgsd}
\title{Weighted geometric standard deviation}
\usage{
wgsd(x, w)
}
\arguments{
\item{x}{a numeric vector.}

\item{w}{a vector of numeric weights of \code{length(x)}.}
}
\value{
Return the weighted geometric standard deviation
of a numeric vector.
}
\description{
Take a numeric vector
and return its weighted geometric standard deviation.
}
\details{
None.
}
\seealso{
\code{\link{wasd}} returns arithmetic standard deviation. \cr
\code{\link{wasd}} returns weighted arithmetic standard deviation. \cr
\code{\link{gsd}}  returns geometric standard deviation. \cr
\code{\link{wgsd}} returns weighted geometric standard deviation.
}
\keyword{internal}
